FEAT: Added Star Ratings View
Created a custom RatingsView that will allow the user to provide a max ratings,
a bound currentRating and a width to determine the width of the view to be displayed and the size of the stars.
The RatingsView will display stars for the max rating which includes a clear star along with a number of stars in an HStack to represent from 0.. ‹maxRating
Tapping on one of the starts will set the current rating.